RedesignForCreatingNewSymbol final version but actually the work done by this branch is trying to use continuous type instead of quadrant type eigen vectors to calculate similarity. However it's failed because continuous type is too detail to recognize symbol written in the air.

+++ symbolreg.h
@@ -63,7 +63,7 @@ ctData ctDataRecoverXZFromIntA(ctData &originCT, intArray intA);
 //Compare similarity
 intArray ctDataXZDistance(ctData data);
 intArray* ctDataToIntA(ctData data);
-double getCorrOfAxisWithNoZero( intArray temp, intArray sample, int times );
+double getCorrOfAxisWithNoZero( intArray temp, intArray sample, double times );
 
 double meanOfABSIntA(intArray A);
 double unMatchedPercent(intArray FA, intArray SA);
